etl/postgres_to_es/requirements.txt
elasticsearch==8.5.0
psycopg2-binary==2.9
pydantic==1.10.2
python-dotenv==0.21.0
redis==4.3.4